BDS#31 adapt menu and extract js script, document it
